test(pytest): The tkinter initialization error
- was caused by multiple test files trying to create separate Tk instances in a headless environment
- Moved the robust root fixture to conftest.py - This eliminates code duplication and ensures consistency
- Session-scoped fixture - Prevents multiple Tk instance creation across test sessions
- Robust error handling - Uses contextlib.suppress() for graceful cleanup
- Smart root reuse - Tries to reuse existing root or creates new one safely
