+++ tests/test_frontend_tkinter_directory_selection_integration.py
@@ -10,7 +10,6 @@ SPDX-FileCopyrightText: 2024-2025 Amilcar do Carmo Lucas <amilcar.lucas@iav.de>
 SPDX-License-Identifier: GPL-3.0-or-later
 """
 
-import contextlib
 import os
 import tkinter as tk
 from collections.abc import Callable, Generator
@@ -54,37 +53,6 @@ class WidgetEventTracker:
             self.widget.unbind(event_name)
         self.bindings.clear()
         self.events.clear()
-
-
-@pytest.fixture(scope="session")
-def root() -> Generator[tk.Tk, None, None]:
-    """Create and clean up Tk root window for testing."""
-    # Try to reuse existing root or create new one
-    try:
-        root = tk._default_root  # type: ignore[attr-defined]
-        if root is None:
-            root = tk.Tk()
-    except (AttributeError, tk.TclError):
-        root = tk.Tk()
-
-    root.withdraw()  # Hide the main window during tests
-
-    # Patch the iconphoto method to prevent errors with mock PhotoImage
-    original_iconphoto = root.iconphoto
-
-    def mock_iconphoto(*args, **kwargs) -> None:
-        pass
-
-    root.iconphoto = mock_iconphoto  # type: ignore[method-assign]
-
-    yield root
-
-    # Restore original method and destroy root
-    root.iconphoto = original_iconphoto  # type: ignore[method-assign]
-
-    # Only destroy if we're the last test
-    with contextlib.suppress(tk.TclError):
-        root.quit()  # Close the event loop
 
 
 # pylint: disable=duplicate-code
